requirements.txt
redis
mysql-connector-python
python-dotenv